Carve sessions store + RecentConceptions out of main.tsx + settings-modal
Two surgical extractions: createSessionsStore() pulls the 30-line live-pty-session tracking block out of main.tsx into its own hook, and RecentConceptionsSection moves from settings-modal.tsx into settings-modal-parts/. Net: main.tsx 1372 → 1342, settings-modal.tsx 874 → 814. Design notes (deferred carve targets, closure-coupling rationale) live alongside the project in the conception tree.
